Request reloads through the CLI request-reload.
Also make the request-reload command configurable in the hostdb.ini like

     [interface]
     request_reload_cmd = /path/to/my/request-reload

This makes the use of JEvent to send reload requests to DNS and DHCP
servers totally optional (although still the default). If you want to
use another mechanism, implement it in a command compatible with
request-reload and point it out through configuration.

It also makes us truly not Depend on libjevent-perl, so the Recommends
becomes true.